commit suggested changes
